feat: Phase 54 — migration audit, BytePCM fix, ast-compare improvements
- Fix BytePCM float→byte conversion to match Java truncation semantics:
  `(f * 127.0) as i32 as i8` instead of saturating `as i8` (resolves 2 regressions)
- Add 15 bmson/osu POJO files to ast-compare ignore list (merged into Rust modules)
- Update AGENTS.md: 2940 tests, 0 regressions, Phase 54 audit findings
- Phase 54 audit: 88% of 257 "missing" methods are false positives from
  architectural redesign (inner class→closure, abstract→enum dispatch, etc.)
  ~28 genuine gaps remain, all documented and non-blocking

+++ lr2oraja-rust/crates/beatoraja-audio/src/byte_pcm.rs
@@ -74,7 +74,10 @@ impl BytePCM {
                         pcm[i * 4 + 2],
                         pcm[i * 4 + 3],
                     ]);
-                    s[i] = (f * i8::MAX as f32) as i8 as u8;
+                    // Java: (byte)(pcm.getFloat() * Byte.MAX_VALUE)
+                    // float→int truncates toward zero, int→byte truncates to low 8 bits.
+                    // Rust `as i8` saturates (since 1.45), so go via i32 first.
+                    s[i] = ((f * i8::MAX as f32) as i32 as i8) as u8;
                 }
                 sample = s;
             }
